# apps/backend/public-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Voholabs Studio Public API",
    "version": "1.0.0",
    "description": "Schedule posts, manage media, read analytics, the agent brief, the skills library and the credit wallet of a Voholabs Studio workspace.\n\nAuthenticate with the workspace's API key (Settings, Public API) or an OAuth token (`pos_...`) in the `Authorization` header, without a `Bearer` prefix.\n\nErrors: a refusal that needs wallet credits answers 402 with `{ message, wallet: true, url: \"/wallet\" }`; `url` is relative to the Studio app. Other refusals carry `msg` or `message`.",
    "contact": {
      "name": "Voholabs",
      "url": "https://voholabs.com/docs/api"
    },
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0-only"
    }
  },
...
  "servers": [
    {
      "url": "https://studio.voholabs.com/api/public/v1"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
...
  "paths": {
...
    "/skills": {
      "get": {
        "tags": [
          "Skills"
        ],
        "summary": "List skills",
        "description": "The skills library: step-by-step methods for content jobs, as a short catalog. Opens with the first wallet top-up (402 with `wallet: true` before that); paid plans always pass.",
        "operationId": "listSkills",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillsList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          }
        },
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Only skills with this tag key"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Words in the name, summary or tags"
          }
        ]
      }
    },
    "/skills/{slug}": {
      "get": {
        "tags": [
          "Skills"
        ],
        "summary": "Get a skill",
        "description": "One skill in full, with its instructions.",
        "operationId": "getSkill",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Skill slug"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "The API key, or an OAuth access token starting with pos_"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Most refusals carry `msg` or `message`.",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "msg": {
            "type": "string"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          }
        }
      },
      "WalletRefusal": {
        "type": "object",
        "description": "Something needs wallet credits or a wallet top-up. Open `url` on the Studio app (relative to it, e.g. https://studio.voholabs.com/wallet) to top up.",
        "properties": {
          "statusCode": {
            "type": "integer",
            "const": 402
          },
          "message": {
            "type": "string"
          },
          "msg": {
            "type": "string",
            "description": "Also present on GET /social/{integration}"
          },
          "wallet": {
            "type": "boolean",
            "const": true
          },
          "url": {
            "type": "string",
            "example": "/wallet"
          }
        },
        "required": [
          "message",
          "wallet",
          "url"
        ]
      },
      "PlanRefusal": {
        "type": "object",
        "description": "Something needs a paid plan.",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "SkillsList": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                }
              }
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillSummary"
            }
          }
        }
      },
      "SkillSummary": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "whenToUse": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "usesBrief": {
            "type": "boolean"
          }
        }
      },
      "Skill": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SkillSummary"
          },
          {
            "type": "object",
            "properties": {
              "whatItDoes": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "body": {
                "type": "string",
                "description": "The skill's instructions, in Markdown"
              },
              "version": {
                "type": [
                  "integer",
                  "string"
                ]
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "No API key, or an invalid one",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PaymentRequired": {
        "description": "Needs wallet credits or a top-up (`wallet: true`), or a paid plan",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/WalletRefusal"
                },
                {
                  "$ref": "#/components/schemas/PlanRefusal"
                }
              ]
            }
          }
        }
      },
...
    }
  }
}